[client,android] Extract SessionDialogs
